+++ tests/integration-test-samples/validation/107/valid_numeric.php
@@ -19,5 +19,5 @@ $expectedErrors = [];
 $ares = new Ares($schema);
 
 $this->assertSame(empty($expectedErrors), $ares->validate($data));
-$this->assertEquals($expectedErrors, $ares->getValidationErrors());
+$this->assertEquals($expectedErrors, $ares->getValidationErrors()->getArrayCopy());