+++ tests/integration-test-samples/validation/122/cyclic_custom_type_references.php
@@ -55,5 +55,5 @@ $expectedErrors = [
 $ares = new Ares($schema);
 
 $this->assertSame(empty($expectedErrors), $ares->validate($data));
-$this->assertEquals($expectedErrors, $ares->getValidationErrors());
+$this->assertEquals($expectedErrors, $ares->getValidationErrors()->getArrayCopy());