+++ tests/integration-test-samples/validation/123/unknown_allowed_per_map_inverted.php
@@ -42,5 +42,5 @@ $expectedErrors = [
 $ares = new Ares($schema);
 
 $this->assertSame(empty($expectedErrors), $ares->validate($data, ['allUnknownAllowed' => true]));
-$this->assertEquals($expectedErrors, $ares->getValidationErrors());
+$this->assertEquals($expectedErrors, $ares->getValidationErrors()->getArrayCopy());